temperature_api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
Django==5.2.3
django-cors-headers==4.7.0
django-ratelimit==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
propcache==0.3.2
PyJWT==2.9.0
sqlparse==0.5.3
yarl==1.20.1
